Fixed BMI Calc

+++ AP_Lesson_03/BMICalc.java
@@ -4,14 +4,14 @@ public class BMICalc {
 	public static void main(String[]args) {
 		Scanner keyboard = new Scanner(System.in);
 		System.out.println("What is your height in inches?");
-		int height = keyboard.nextInt();
+		double height = keyboard.nextDouble();
 		
 		System.out.println("What is your weight in pounds?");
-		int weight = keyboard.nextInt();
+		double weight = keyboard.nextDouble();
 		
-		int heightSqr = height * height;
-		float BMIStep2 = weight / heightSqr;
-		float BMI = BMIStep2 * 703;
+		double heightSqr = height * height;
+		double BMIStep2 = weight / heightSqr;
+		double BMI = BMIStep2 * 703;
 		System.out.println("Your BMI is " + BMI + ".");
 	}
 }
